more update.

+++ cleanForNewTest.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
 rm -rf $HOME/.eBlocBroker/*/*
 rm -f  $HOME/.eBlocBroker/my-app.pid
 rm -f  $HOME/.eBlocBroker/checkSinfoOut.txt
@@ -8,7 +13,7 @@ rm -f  $HOME/.eBlocBroker/queuedJobs.txt
 rm -f  $HOME/.eBlocBroker/test.txt
 rm -f  $HOME/.eBlocBroker/ipfs.out
 
-cat /dev/null > clusterDriver.out
+cat /dev/null > $HOME/.eBlocBroker/clusterDriver.out
 
 sudo bash killall.sh
 sudo bash clean.sh
